Move HistogramImpl to a separate file
